segundo commit

+++ crud-cassandra/src/main/java/br/edu/ifpb/monteiro/ads/dao/AlunoDao.java
@@ -21,7 +21,7 @@ public class AlunoDao {
 
 	private Conexao conexao;
 	private Session session;
-	private ValidacaoAluno valida;
+	private ManipuladorAluno manipulador = new ManipuladorAluno(this);
 
 	public AlunoDao(Conexao conexao) {
 		this.conexao = conexao;
@@ -31,8 +31,7 @@ public class AlunoDao {
 	public void adicionar(Aluno aluno) throws MatriculaInvalidaException,
 			MatriculaDuplicadaException, NomeInvalidoException,
 			CursoInvalidoException, TelefoneInvalidoException,
-			CpfDuplicadoException, CpfInvalidoException, DadoInvalidoException,
-			AlunoInexistenteException {
+			CpfDuplicadoException, CpfInvalidoException, DadoInvalidoException {
 		if ((aluno.getCpf() == null || aluno.getCpf().equals("   .   .   -  "))
 				&& (aluno.getNome() == null || aluno.getNome().trim()
 						.equals(""))
@@ -44,8 +43,8 @@ public class AlunoDao {
 						.equals("            "))) {
 			throw new DadoInvalidoException();
 		} else {
-			valida = new ValidacaoAluno(this);
-			valida.validar(aluno, null, null);
+			manipulador = new ManipuladorAluno(this);
+			manipulador.validar(aluno, null, null);
 
 			session.execute("INSERT INTO cadastro.aluno (matricula, curso, nome, cpf, telefone)"
 					+ "VALUES "
@@ -61,15 +60,19 @@ public class AlunoDao {
 	}
 
 	public void deletar(Aluno aluno) throws AlunoInexistenteException {
-		session.execute("DELETE FROM cadastro.aluno WHERE matricula='"
-				+ aluno.getMatricula() + "';");
+		if (manipulador.buscarMatricula(aluno.getMatricula()) == null) {
+			throw new AlunoInexistenteException();
+		} else {
+			session.execute("DELETE FROM cadastro.aluno WHERE matricula='"
+					+ aluno.getMatricula() + "';");
+		}
 	}
 
 	public void alterar(Aluno aluno, String matricula, String cpf)
-			throws AlunoInexistenteException, MatriculaDuplicadaException,
-			MatriculaInvalidaException, NomeInvalidoException,
-			CursoInvalidoException, TelefoneInvalidoException,
-			DadoInvalidoException, CpfInvalidoException, CpfDuplicadoException {
+			throws MatriculaDuplicadaException, MatriculaInvalidaException,
+			NomeInvalidoException, CursoInvalidoException,
+			TelefoneInvalidoException, DadoInvalidoException,
+			CpfInvalidoException, CpfDuplicadoException {
 		if ((aluno.getCpf() == null || aluno.getCpf().equals("   .   .   -  "))
 				&& (aluno.getNome() == null || aluno.getNome().trim()
 						.equals(""))
@@ -81,43 +84,46 @@ public class AlunoDao {
 						.equals("            "))) {
 			throw new DadoInvalidoException();
 		} else {
-			valida = new ValidacaoAluno(this);
-			valida.validar(aluno, cpf, matricula);
+			manipulador.validar(aluno, cpf, matricula);
 			session.execute("UPDATE cadastro.aluno SET nome='"
 					+ aluno.getNome() + "', cpf='" + aluno.getCpf()
-					+ "', curso='" + aluno.getCurso()
-					+ "', telefone='" + aluno.getTelefone()
-					+ "' WHERE matricula='" + aluno.getMatricula() + "';");
+					+ "', curso='" + aluno.getCurso() + "', telefone='"
+					+ aluno.getTelefone() + "' WHERE matricula='"
+					+ aluno.getMatricula() + "';");
 		}
 	}
 
 	public Aluno consultar(String matricula) throws AlunoInexistenteException {
-		ResultSet results = session
-				.execute("select * from cadastro.aluno where matricula='"
-						+ matricula + "';");
-		String nome = "";
-		String matri = "";
-		String curso = "";
-		String cpf = "";
-		String telefone = "";
-		for (Row row : results) {
-			nome = row.getString("nome");
-			matri = row.getString("matricula");
-			curso = row.getString("curso");
-			cpf = row.getString("cpf");
-			telefone = row.getString("telefone");
-		}
-		Aluno aluno = new Aluno();
-		aluno.setNome(nome);
-		aluno.setMatricula(matri);
-		aluno.setCurso(curso);
-		aluno.setCpf(cpf);
-		aluno.setTelefone(telefone);
+		if (manipulador.buscarMatricula(matricula) != null && manipulador.buscarMatricula(matricula).getMatricula().equals(matricula)) {
+			ResultSet results = session
+					.execute("select * from cadastro.aluno where matricula='"
+							+ matricula + "';");
+			String nome = "";
+			String matri = "";
+			String curso = "";
+			String cpf = "";
+			String telefone = "";
+			for (Row row : results) {
+				nome = row.getString("nome");
+				matri = row.getString("matricula");
+				curso = row.getString("curso");
+				cpf = row.getString("cpf");
+				telefone = row.getString("telefone");
+			}
+			Aluno aluno = new Aluno();
+			aluno.setNome(nome);
+			aluno.setMatricula(matri);
+			aluno.setCurso(curso);
+			aluno.setCpf(cpf);
+			aluno.setTelefone(telefone);
 
-		return aluno;
+			return aluno;
+		} else{
+			throw new AlunoInexistenteException();
+		}
 	}
 
-	public ArrayList<Aluno> listar() throws AlunoInexistenteException {
+	public ArrayList<Aluno> listar() {
 		ResultSet results = session.execute("select * from cadastro.aluno;");
 		String nome = "";
 		String matri = "";
@@ -139,7 +145,6 @@ public class AlunoDao {
 			aluno.setTelefone(telefone);
 			alunos.add(aluno);
 		}
-
 		return alunos;
 	}
